+++ Forms/CropForm.vb
@@ -12,6 +12,7 @@ Public Class CropForm
                 components.Dispose()
             End If
         End If
+        Close()
         MyBase.Dispose(disposing)
     End Sub