+++ General/General.vb
@@ -830,7 +830,8 @@ table {
             IsClosed = True
 
             Writer.WriteRaw("<p>&nbsp;</p>" + BR)
-            Writer.WriteRaw("<h5 align=""center"">Copyright &copy; 2002-" & DateTime.Now.Year & " stax76. All rights reserved.</h5><br>")
+            Writer.WriteRaw("<h5 align=""center"">Copyright &copy; 2017-" & DateTime.Now.Year & " Revan654. All rights reserved.</h5><br>")
+            Writer.WriteRaw("<h5 align=""center"">Copyright &copy; 2002-2017" & " stax76. All rights reserved.</h5><br>")
             Writer.WriteEndElement() 'body
             Writer.WriteEndElement() 'html
             Writer.Close()